image-ination/requirements-inference.txt
tensorflow==1.14
sklearn==0.0
keras==2.3.1
seaborn==0.10.0
matplotlib==3.2.1
pillow==7.0.0
opencv-python==4.2.0.32
scikit-image==0.16.2
IPython==7.13.0
waitress==1.4.3
click==7.1.1
Flask-API==2.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Werkzeug==0.16.1
h5py==2.10.0
